[CartBundle] Cart repository has been moved to component
