+++ src/Sylius/Bundle/CartBundle/Purger/ExpiredCartsPurger.php
@@ -12,7 +12,7 @@
 namespace Sylius\Bundle\CartBundle\Purger;
 
 use Doctrine\Common\Persistence\ObjectManager;
-use Sylius\Bundle\CartBundle\Repository\CartRepositoryInterface;
+use Sylius\Component\Cart\Repository\CartRepositoryInterface;
 use Sylius\Component\Cart\Model\CartInterface;
 use Sylius\Component\Cart\Purger\PurgerInterface;
