+++ src/Sylius/Component/Cart/Repository/CartRepositoryInterface.php
@@ -9,8 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\CartBundle\Repository;
+namespace Sylius\Component\Cart\Repository;
 
+use Sylius\Component\Cart\Model\CartInterface;
 use Sylius\Component\Order\Repository\OrderRepositoryInterface;
 
 /**
@@ -23,7 +24,7 @@ interface CartRepositoryInterface extends OrderRepositoryInterface
     /**
      * Get expired carts
      *
-     * @return array
+     * @return CartInterface[]
      */
     public function findExpiredCarts();
 }
